+++ src/main/java/com/nbrichau/vanillaextension/fences/CoarseDirtFence.java
@@ -1,6 +1,6 @@
 package com.nbrichau.vanillaextension.fences;
 
-import com.nbrichau.vanillaextension.init.BlockInit;
+import com.nbrichau.vanillaextension.init.FenceInit;
 import net.minecraft.block.BlockState;
 import net.minecraft.block.FenceBlock;
 import net.minecraft.entity.player.PlayerEntity;
@@ -22,7 +22,7 @@ public class CoarseDirtFence extends FenceBlock {
 	public ActionResultType onBlockActivated(BlockState state, World worldIn, BlockPos pos, PlayerEntity player, Hand handIn, BlockRayTraceResult hit) {
 		if (!worldIn.isRemote()) {
 			if (player.getHeldItem(handIn).getToolTypes().contains(ToolType.HOE)) {
-				BlockState bs = BlockInit.dirt_fence.getDefaultState().with(NORTH, state.get(NORTH)).with(EAST, state.get(EAST)).with(SOUTH, state.get(SOUTH)).with(WEST, state.get(WEST)).with(WATERLOGGED, state.get(WATERLOGGED));
+				BlockState bs = FenceInit.dirt_fence.getDefaultState().with(NORTH, state.get(NORTH)).with(EAST, state.get(EAST)).with(SOUTH, state.get(SOUTH)).with(WEST, state.get(WEST)).with(WATERLOGGED, state.get(WATERLOGGED));
 				worldIn.setBlockState(pos, bs, 11);
 				worldIn.playSound(null,pos, SoundEvents.ITEM_HOE_TILL, SoundCategory.BLOCKS, 1.0F,1.0F);
 				player.getHeldItem(handIn).damageItem(1, player, item->item.sendBreakAnimation(handIn));