+++ src/main/java/com/nbrichau/vanillaextension/slabs/CoarseDirtSlab.java
@@ -1,6 +1,6 @@
 package com.nbrichau.vanillaextension.slabs;
 
-import com.nbrichau.vanillaextension.init.BlockInit;
+import com.nbrichau.vanillaextension.init.SlabInit;
 import net.minecraft.block.BlockState;
 import net.minecraft.block.SlabBlock;
 import net.minecraft.entity.player.PlayerEntity;
@@ -22,7 +22,7 @@ public class CoarseDirtSlab extends SlabBlock {
 	public ActionResultType onBlockActivated(BlockState state, World worldIn, BlockPos pos, PlayerEntity player, Hand handIn, BlockRayTraceResult hit) {
 		if (!worldIn.isRemote()) {
 			if (player.getHeldItem(handIn).getToolTypes().contains(ToolType.HOE)) {
-				BlockState bs = BlockInit.dirt_slab.getDefaultState().with(TYPE, state.get(TYPE)).with(WATERLOGGED, state.get(WATERLOGGED));
+				BlockState bs = SlabInit.dirt_slab.getDefaultState().with(TYPE, state.get(TYPE)).with(WATERLOGGED, state.get(WATERLOGGED));
 				worldIn.setBlockState(pos, bs, 11);
 				worldIn.playSound(null,pos, SoundEvents.ITEM_HOE_TILL, SoundCategory.BLOCKS, 1.0F,1.0F);
 				player.getHeldItem(handIn).damageItem(1, player, item->item.sendBreakAnimation(handIn));