+++ changelog.md
@@ -5,10 +5,12 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ## [Unreleased]
 ### Added
+- Trapdoor variant for most of the vanilla blocks.
 - Stairs, slabs, fences and walls variant for leaves block, vine and grass block use overlay textures. The color change depending of the biome.
 - Log, Wood, Stem and Hyphae can be stripped to get the stripped version of the block. Work with stairs, slabs, fences and walls.
 - Secret advancements
 - You can convert Coarse Dirt Blocks to Dirt Blocks with a Hoe.
+- VanillaExtension Grass Blocks can spread to all Dirt Blocks. (Minecraft Grass Block can't spread on VanillaExtension Grass Block )
 
 ### Changed
 - recipes for glass walls don't conflict with glass pane recipe anymore. Recipe for these walls are like the recipes for plank walls